==> nbs/00_README_and_Data_Exploration.ipynb <==
# !python --version

# OnDeck ML Challenge

This notebook is part of the [OnDeck ML Challenge](https://gist.github.com/thaumant/ea2f03b5546f36b72443357238edd2d9)

- Build the first iteration of a recommender system that could suggest top-recommended movies to users.
- Provide a GitHub repo (private is fine) with the working model and some description of how to reproduce training/testing steps.
- Provide written answers to these questions:
  - A brief justification of the chosen model and the scoring metric. What alternatives would be worth exploring?
  - What would be your next steps to improve the recommendation quality?

Note: I've left all outputs inline, and you're not required to run this code. To the same effect, I'm not adding a requirements.txt

## Chosen Modeling Approaches

1. SVD
2. Content Based with Nearest Neighobur Recommendation
3. Neural Collaborative Filtering 

Whenever possible, I use the implicit feedback formulation of the problem.

## Scoring Metric

For any recommendation system, we've multiple choice of a scoring or evaluation metric. We've search-inspired metrics: Hit Ratio, Precision @K, Recall @K, mean reciprocal rank, nDCG and so on. We also have regression-inspired metrics e.g. RMSE, MAE.

I'm picking metrics with 2 separate intents: 
1. Being able to compare multiple approaches, independent of approach
2. Modeling specific loss metrics (and evaluation) when needed

### Hit Ratio @10 - Implicit Feedback

In modern contexts, a lot of movie recommendations happen in the context of online viewing e.g. Netflix or Prime Video. At any given time, they need only "one" correct from their entire panel of recommendations. 

Based on this context, Hit Rate @10 is a preferred metric in this context. To simplify calculation, we count all movies which have a rating >= threshold as "seen". This is also the metric implemented across both implementations. 

Additionally, I use RMSE to evaluate SVD, and a Binary Cross Entropy Loss for Neural CF training.

## Improving Recommendation Quality

There are 2 main approaches for improving recommendation quality

**Modeling/Algorithmic Improvements**

1. In addition to SVD, other matrix factorisation methods can be looked into e.g. [ALS](https://www.elenacuoco.com/2016/12/22/alternating-least-squares-als-spark-ml/#:~:text=ALS%20recommender%20is%20a%20matrix,algorithm%20in%20a%20parallel%20fashion.) is a well-known and popular baseline which we can look into

1. Parameter turning for all the approaches: While I've used a **GridSearchCV** based approach for SVD optimisation - I didn't optimise the hyperparameters for Neural Collaborative Filtering

1. In content-based system, using a Categorical Encoder for Genres and Years can be a viable alternative to simply absorbing it as a part of the movie description.

1. Better/Domain-specific Embedding: We use embedding which are trained on a large open-domain dataset ranging from Wikipedia to Reddit. Maybe, a movies specific embedding built using a Language Model trained on it would yield better similarity results.

1. Lastly, given the relatively small amount of training data - it might be effective to explore an approach which blends content and collaborative filtering based approaches. 
    - A straightforward to do so would be to use content based system to generate candidates and then use a collaborative system to rank and select which to recommend.
   

**Improve Dataset Usage**

1. In the present implementation, I do not use the `teams` or `aggs` information at all. Using `teams`, we can exploit actor/writer/director preferences of the user, if any.

1. Incorporate popular demographic information (from `aggs`) and synopsis (from `movies`) to improve the similarity recommendation
    - To use synopsis in content recommendation system, we will use a different text encoder which can do a better job at long lengths than a sentence. Specifically, we might choose something like a Pretrained [Longformer](https://huggingface.co/transformers/model_doc/longformer.html).
    - In order to illustrate how this might look in practice, I've implemented (and tested) a class `SynopsisProc` based on this idea, but using [BERT](https://huggingface.co/bert-base-uncased) because of it's popularity.
    - We would then combine (e.g. concatenate) with our movie `description` attribute to recommend movies

1. Given any information about the user itself, e.g. the demographic - we can establish a relationship between the user and movie's popular demographic.

[Dataset link](https://www.dropbox.com/s/vi7lktdxx0r97o4/od-challenge.tar.gz?dl=1).

## Notes

- You define the technical details of the solution: the model, feature set, test/train split, normalization, loss, scoring metric.
- You're not expected to use as much features as possible. Explore the data and take what works best for your approach.
- You're not expected to produce a model with perfect scores. It's enough to pick a sensible model, make a few tweak iterations, produce some results, and outline a path to improve it.

## Technical constraints

- Python ~3.6~ 3.8+.
- The model, the traing and testing code should be in Jupyter notebooks. Everything else may be in notebooks or python files.
- If trainng takes more than an hour on a laptop, provide the trained model and the code to load it.

## Dataset details

`movies.pickle` (4107 rows) — basic info about movies:

| Column     | Type       | Example                             | Notes                     |
| :--------- | :--------- | :---------------------------------- | :------------------------ |
| `movie_id` | `int`      | `109830`                            |
| `title`    | `str`      | `"Forrest Gump"`                    |
| `genres`   | `set[str]` | `{"Romance", "Comedy"}`             |
| `year`     | `int`      | `1994`                              |
| `synopsis` | `str`      | `"The film begins with feather..."` | Detailed plot description |

`aggs.pickle` (28557 rows) — aggregated ratings, total and by demographic:

| Column           | Type    | Example       | Notes                                     |
| :--------------- | :------ | :------------ | :---------------------------------------- |
| `movie_id`       | `int`   | `109830`      |
| `rating_average` | `float` | `8.8`         |
| `rating_count`   | `int`   | `304`         | Number of ratings collected for the group |
| `demographic`    | `str`   | `"age_18_29"` | Group name: total, by age, by gender      |

`teams.pickle` (190547 rows) — cast & crew:

| Column        | Type  | Example             | Notes                       |
| :------------ | :---- | :------------------ | :-------------------------- |
| `movie_id`    | `int` | `109830`            |
| `person_role` | `str` | `"actor"`           | Enum: actor/director/writer |
| `person_id`   | `int` | `37097`             |
| `person_name` | `str` | `"Giovanni Arpino"` |

`labels.pickle` (42237 rows) — user ratings to use as labels:

| Column     | Type    | Example  | Notes      |
| :--------- | :------ | :------- | :--------- |
| `movie_id` | `int`   | `109830` |
| `user_id`  | `int`   | `184`    |
| `rating`   | `float` | `8.8`    | 1-10 scale |

import pickle
import random
import tarfile
import urllib.request
from pathlib import Path
from typing import List

import autoreload
import pandas as pd
from sklearn.model_selection import train_test_split

%load_ext autoreload
%autoreload 2

Path.ls = lambda x: list(x.iterdir())

def get_data(
    url: str = "https://www.dropbox.com/s/vi7lktdxx0r97o4/od-challenge.tar.gz?dl=1",
    file_name: str = "../data/raw/od-challenge.tar.gz",
):
    # Download the file from `url` and save it locally under `file_name`:
    urllib.request.urlretrieve(url, file_name)
    return None


get_data()

with tarfile.open("../data/raw/od-challenge.tar.gz") as file:
    file.extractall("../data/ext")

data_dir = Path("../data/ext/od-challenge")
assert data_dir.exists()
files = data_dir.ls()
files

def read(data_dir, ds: str):
    with (data_dir / f"{ds}.pickle").open("rb") as f:
        df = pickle.load(f)
    return df

aggs = read(data_dir, "aggs")
teams = read(data_dir, "teams")
movies = read(data_dir, "movies")
labels = read(data_dir, "labels")

print(
    f"There are {labels.movie_id.nunique()} movies with {len(labels)} ratings against them. There are also {labels.user_id.nunique()} users only"
)

The relatively small number of user-movie ratings indicates that training a deep learning model from scratch might not be a great approach. We still have the option of using a deep learning based approaches, which don't require large amounts of training data.

labels.head()
labels.movie_id.nunique(), labels.user_id.nunique(), labels.rating.value_counts()

movies.head()
movies.movie_id.nunique(), movies.genres.value_counts(), movies.year.value_counts()

teams.head()
teams.movie_id.nunique(), teams.person_role.nunique(), teams.person_id.nunique(), teams.person_id.value_counts()

genres = movies.genres.tolist()

flat_genres = []
for g in genres:
    flat_genres.extend(list(g))

from collections import Counter

Counter(flat_genres)

aggs.head()

aggs.movie_id.nunique()

# Possible Features for Every Movie:

1. Genre (Category Encoded?)
2. Year of Release (Cont.)
3. Vector Emb of Synopsis, say 786 or 300 dim
4. Some combination of Top_Demo * Rating_Count of Ratings Average -- captures information about "Popular in" what demo
5. Preference in actor/writer/director? What about person_id? There's too many values there. Maybe retain the most popular? 

# How to split data?

There is more than one way to think about recommending a movie. Here are few: 

1. Leave-one-last-out: Leave one last movie from a stream of user rated movies, and predict the next movie
2. Random Movie: For each user, split interactions into train and test data
3. Random User: Split some users into train, the rest into test
4. Random Rating: Split some ratings into train, rest into test - without ensuring that a user or a movie is absent is leaking in test or not
5. Time-based: Split within each user, based on rating timestamp

Since we don't have have time of rating, we can safely ignore that. Last one out also requires making some assumption about the data ordering, so I am choosing to skip that. From the remaining options - to keep things simple, I'm choosing `Random Rating`. 

# Simple Baseline: Popular Movies


Before we start modeling the movie-user interaction and preference, I'd like to implement a baseline approach: 
What if recommend the most popular Top 10 movies to every user? When measuring popularity, we're looking for movies which are rated often - independent of their rating. We limit ourselves to **Top 10** movies. 

labels.describe()

train, test = train_test_split(
    labels, test_size=0.2, shuffle=True, stratify=labels.rating, random_state=42
)

target_dir = Path("../data/intermediate")
target_dir.mkdir(exist_ok=True)
train.to_csv("../data/intermediate/train.csv", index=False)
test.to_csv("../data/intermediate/test.csv", index=False)

!ls -l {target_dir}

top_k = 10
vc = dict(train.movie_id.value_counts())
popular_movies = [k for k, v in vc.items()][:top_k]
popular_movies

# How do we count a hit? 

1. We predict a list of Top K movies which the user should've rated
2. If the user has infact seen ANY of the Top K moves, we get a hit, else a miss
3. A mean across all users in target, is our hit rate

def user_hits(predicted_movies: List[int], seen_movies: List[int]):
    return len(set(predicted_movies) & set(seen_movies)) > 0


def calc_hit_rate(split):
    hits = []
    for user_id in split.user_id:
        seen_movies = split[split.user_id == user_id].movie_id.unique()
        hits.append(user_hits(popular_movies, seen_movies))

    return sum(hits) / len(hits)


calc_hit_rate(test)
