==> nbs/02_Content_Based_RecSys.ipynb <==
# Alternative Approach: Content Based kNN Recommendation

## Why? 

Content Based recommendation is excellent at solving the cold start problem, if there are not too many movies seen by a particular user. This often has a different recommendation compared to Collaborative Filtering approaches, and that adds to the quality of the final recommendation when we can blend the two in some meaningful way. 

# Implementation

## Part 1: Similar Movies

We find the most similar movies, given a single input movie:

1. Merge into single sentence: The movie information has a lot of text and numerical information. I combined the movie title, year of release and genre into a single sentence. 
2. Vectorize: The present state of the art, which builds on BERT, but is finetuned on over a billion sentence pairs: `sentence-transformers/all-mpnet-base-v2`
3. Annoy Index: [Annoy](https://github.com/spotify/annoy) allows me to do fast approximate nearest neighbour queries using both: item and vector.

We could also use a similar approach for Synopsis, but given the limited time - I've only illusrated that with code towards the end. 

## Part 2: Recommend Movies to User

1. For each movie which is seen by the user, we pull the k (=3) Nearest Neighbours here which are unseen. 
2. Aggregate and sort by count and most similar
3. Return the Top 10

# !pip install --upgrade torch

# !pip install sentence-transformers -qq

import pickle
import random
from pathlib import Path
from typing import List

import numpy as np
import pandas as pd
import torch
from annoy import AnnoyIndex
from sentence_transformers import SentenceTransformer
from sklearn.metrics import mean_squared_error
from tqdm import tqdm
from transformers import AutoModel, AutoTokenizer, BertModel, BertTokenizer

tqdm.pandas()

def read(ds: str, data_dir=Path("../data/ext/od-challenge")):
    with (data_dir / f"{ds}.pickle").open("rb") as f:
        df = pickle.load(f)
    return df


aggs = read(ds="aggs")
teams = read(ds="teams")
movies = read(ds="movies")
labels = read(ds="labels")

data_dir = Path("../data/intermediate/")
train, test = pd.read_csv(data_dir / "train.csv"), pd.read_csv(data_dir / "test.csv")

# def write(df, ds, data_dir=Path("../data/intermediate")):
#     df.to_csv(data_dir / f"{ds}.csv", index=False)


# write(aggs, ds="aggs")
# write(teams, ds="teams")
# write(movies, ds="movies")
# write(labels, ds="labels")

df = movies
df

description, genre_text = [], []
for row in df.iterrows():
    info = row[1]
    genres = list(info["genres"])
    genres = ", ".join(genres)
    text = f"Movie Title: **{info['title']}** was released in {info['year']}"
    genres = f"Genres: {genres}"
    description.append(text)

df["description"] = pd.Series(description)

model = SentenceTransformer('sentence-transformers/all-mpnet-base-v2')

def get_description_index(new=False):
    vector_dir = Path("../models/vector_index/")
    fname = "description.ann"
    index_path = vector_dir / fname
    sz = 768
    if index_path.exists() and not new:
        u = AnnoyIndex(sz, "angular")
        u.load(str(index_path))  # super fast, will just mmap the file
        return u
    else:
        embeddings = df.description.progress_apply(lambda x: model.encode(x))
        # takes about 20 minutes to run locally on an old 2015 Mac Air

        t = AnnoyIndex(sz, "angular")  # Length of item vector that will be indexed
        for i, vector in enumerate(embeddings):
            t.add_item(i, vector)
        t.build(len(embeddings) // 10)
        t.save(str(index_path))
        return embeddings, t


embeddings, t = get_description_index(new=True)

# Demo: Movie-Movie Recommendation

random.seed(37)
idx = random.choice(movies.movie_id.tolist())


def similar_movies(movie_id):
    idx = movies[movies.movie_id == movie_id].index[0]
    indices, distances = t.get_nns_by_item(idx, 4, include_distances=True)
    return movies.loc[idx]["description"], movies.loc[indices][1:], distances[1:]


input_movie, recommended_movies, distances = similar_movies(idx)
print(recommended_movies.movie_id, distances)
# print(input_movie, "\n---------\n", recommended_movies)

Next, we use the movie-movie recommendation to recommend movies to our user - as per the original statement. 
In order to do this, we find movies similar to the movies the user has already seen and recommend those. 

The assumption is that there all users have liked (rating >= 3.5) atleast 2 movies, 1 each across train and test split. 

**Note on Hit Ratio**:
> A hit is counted if any of the movies we recommend, is present in the movies the user rates in test.

set(test.user_id.unique()) - set(train.user_id.unique())

def user_hits(predicted_movies: List[int], seen_movies: List[int]):
    return len(set(predicted_movies) & set(seen_movies)) > 0


def get_seen_movies(test, user_id, threshold=3.0):
    df = test[test.user_id == user_id]
    seen_movies = []
    for user_item_rating in df.iterrows():
        if user_item_rating[1]["rating"] >= threshold:
            seen_movies.append(user_item_rating[1]["movie_id"])
    return seen_movies


def recommend_movies(user_id, train, threshold=3.0, k=10):
    train = train[train.user_id == user_id]
#     train = train[train.rating >= threshold]
    recommended_movies = []
    for movie_id in train.movie_id:
        _, movie_df, distances = similar_movies(movie_id)
        movie_df["similarity"] = [1 - d for d in distances]
        recommended_movies.append(movie_df)
    try:
        recommended_movies = pd.concat(recommended_movies)
    except ValueError:
        return []

    seen_movies = train.movie_id.unique()
    recommended_movies["seen"] = recommended_movies.movie_id.apply(
        lambda x: x not in seen_movies
    )
    recommended_movies = recommended_movies[recommended_movies.seen]
    
    
    recommended_movies["counts"] = recommended_movies["movie_id"].map(
        recommended_movies["movie_id"].value_counts()
    )
    recommended_movies.sort_values(
        by=["similarity"], inplace=True, ascending=False
    )
    return recommended_movies.movie_id.unique()[:k]


def calc_hit_rate(split, train):
    hits = []
    for user_id in tqdm(split.user_id.unique()):
        recommended_movies = recommend_movies(user_id=user_id, train=train)
#         for mv in recommended_movies:
#             print(movies[movies.movie_id == mv][["title", "year", "genres"]])
        seen_movies = get_seen_movies(split, user_id)

#         print("-----------------")
#         for mv in seen_movies:
#             print(movies[movies.movie_id==mv][["title", "genres"]])
#         print("##############")
        hits.append(
            user_hits(predicted_movies=recommended_movies, seen_movies=seen_movies)
        )

    return sum(hits) / len(hits)


calc_hit_rate(test, train)

## Synopsis Embedding Illustration

# class SynopsisProc:
#     def __init__(self):
#         self.tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
#         self.model = BertModel.from_pretrained("bert-base-uncased")

#     def embed(self, paragraphs):
#         encoded_input = self.tokenizer(
#             paragraphs, padding=True, truncation=True, return_tensors="pt"
#         )
#         output = self.model(**encoded_input)
#         embedding = output.pooler_output
#         return embedding

# sp = SynopsisProc()

# synopsis_embedding = df['synopsis'].progress_apply(lambda x: sp.embed(x))
